integration testing of comment, label, instruction, linestmt, linestmtseq and node classes
