+++ src/edu/soen341/projectb/tests/unit/Test_Instruction_inherent.java
@@ -1,14 +1,14 @@
-package edu.soen341.projectb.tests;
+package edu.soen341.projectb.tests.unit;
 
 import edu.soen341.projectb.assembler.Instruction;
 
-public class Test_Instruction_immediate {
+public class Test_Instruction_inherent {
     public static void main(String[] args){
         Instruction instruction = new Instruction();
-        System.out.print("Test immediate instruction\n");
-        System.out.print(String.format("0x48 br.15"));
+        System.out.print("Test inherent instruction\n");
+        System.out.print(String.format("0x01 pop"));
 
-        instruction.setInstruction("br.i5");
+        instruction.setInstruction("pop");
         System.out.println("\n0x" + instruction.printInstruction()[0] + " " +
                 instruction.printInstruction()[1]);
     }
