+++ src/edu/soen341/projectb/tests/unit/Test_SymbolTable.java
@@ -1,4 +1,4 @@
-package edu.soen341.projectb.tests;
+package edu.soen341.projectb.tests.unit;
 
 import edu.soen341.projectb.nodes.SymbolTable;
